Started splitting Patient and Encounter model and CRUD code.
